coef.dmr fix
Changes to the Matrix package in version 1.3-1 causes the command Matrix(0) to now return a ddiMatrix diagonal matrix instead of a dsCMatrix. This leads to an error when computing DMR coefficients of a counts matrix that has multiple columns with no observations. Adding the argument doDiag = FALSE creates a dsCMatrix and solves the issue.